Minimal version
implemented the backend fetching functions into the "protoscrat.py" Module

Copied the required methods (sendData, setCaption, clearCreatedInputs), and linked user input ("UserId") to fetch associated posts (up to 40)

Removed old drafts py files
